move zod schema to shared lib
